Refactored and organized code for production readiness
